+++ services/users/project/tests/test_users.py
@@ -1,15 +1,7 @@
 import json
 import unittest
-from project import db
-from project.api.models import User
 from project.tests.base import BaseTestCase
-
-
-def add_user(username, email):
-    user = User(username=username, email=email)
-    db.session.add(user)
-    db.session.commit()
-    return user
+from project.tests.utils import add_user
 
 
 class TestUserService(BaseTestCase):