+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Flask-React ChangeLog
 
+## 0.1.2 2018-11-06
+- Implemented `flask-migrate`
+- `User.email` max length 254 and Unique
+- `User.username` max length 32 and Unique
+
 ## 0.1.1 2018-11-06
 - Moved github template files from root level to `/.github`